Add flake.nix, build with nix by default, update resolver
Interestingly, two constraints are now solveable with less fuel, causing two test failures:

```console
Tests
  Unit
    horn-neg-el
      ple_sum_fuel.5.smt2: FAIL (0.04s)
        tests/test.hs:118:
        Wrong exit code
        expected: ExitFailure 1
         but got: ExitSuccess
        Use -p '/ple_sum_fuel/&&/horn-neg-el.ple_sum_fuel.5.smt2/' to rerun this test only.
    horn-neg-na
      ple_sum_fuel.5.smt2: FAIL (0.03s)
        tests/test.hs:118:
        Wrong exit code
        expected: ExitFailure 1
         but got: ExitSuccess
        Use -p '/ple_sum_fuel/&&/horn-neg-na.ple_sum_fuel.5.smt2/' to rerun this test only.
```

Lowering the fuel parameter in the negative case to 3 (from 5) is the most we can give it before it succeeds.

Lowering the fuel parameter in the positive case to 4 (from 6) is the least we can give it before it fails.